requirements.txt
ConfigParser==6.0.0
designate==16.0.1
docutils==0.20.1
fixtures==4.1.0
importlib_metadata==6.8.0
keystoneauth1==5.2.1
mock==5.1.0
nose==1.3.7
oslo.config==9.1.1
oslo_log==5.2.0
pynetbox==7.0.1
python_novaclient==18.3.0
reno==4.0.0
setuptools==56.0.0
six==1.16.0
Sphinx==7.1.2
testrepository==0.0.20
testresources==2.0.1
testscenarios==0.5.0
testtools==2.6.0
virtualenv==20.24.3
wheel==0.41.1
